3.3.6

+++ shadowsocks-csharp/Controller/PolipoRunner.cs
@@ -8,21 +8,28 @@ using System.IO.Compression;
 using System.Text;
 using System.Net.NetworkInformation;
 using System.Net;
+using System.Runtime.InteropServices;
 
 namespace Shadowsocks.Controller
 {
     class PolipoRunner
     {
         private Process _process;
-        private static string temppath;
+        private static string runningPath;
         private int _runningPort;
 
         static PolipoRunner()
         {
-            temppath = Path.GetTempPath();
+
+            runningPath = Path.Combine(System.Windows.Forms.Application.StartupPath, @"temp"); // Path.GetTempPath();
+            if (!Directory.Exists(runningPath))
+            {
+                Directory.CreateDirectory(runningPath);
+            }
             try
             {
-                FileManager.UncompressFile(temppath + "/ss_polipo.exe", Resources.polipo_exe);
+                FileManager.UncompressFile(runningPath + "/ss_privoxy.exe", Resources.privoxy_exe);
+                FileManager.UncompressFile(runningPath + "/mgwz.dll", Resources.mgwz_dll);
             }
             catch (IOException e)
             {
@@ -43,30 +50,42 @@ namespace Shadowsocks.Controller
             Server server = configuration.GetCurrentServer();
             if (_process == null)
             {
-                Process[] existingPolipo = Process.GetProcessesByName("ss_polipo");
+                Process[] existingPolipo = Process.GetProcessesByName("ss_privoxy");
                 foreach (Process p in existingPolipo)
                 {
+                    string str;
                     try
                     {
-                        p.Kill();
-                        p.WaitForExit();
+                        str = p.MainModule.FileName;
                     }
-                    catch (Exception e)
+                    catch (Exception)
                     {
-                        Console.WriteLine(e.ToString());
+                        continue;
+                    }
+                    if (str == Path.GetFullPath(runningPath + "/ss_privoxy.exe"))
+                    {
+                        try
+                        {
+                            p.Kill();
+                            p.WaitForExit();
+                        }
+                        catch (Exception e)
+                        {
+                            Console.WriteLine(e.ToString());
+                        }
                     }
                 }
-                string polipoConfig = Resources.polipo_config;
+                string polipoConfig = Resources.privoxy_conf;
                 _runningPort = this.GetFreePort();
                 polipoConfig = polipoConfig.Replace("__SOCKS_PORT__", configuration.localPort.ToString());
                 polipoConfig = polipoConfig.Replace("__POLIPO_BIND_PORT__", _runningPort.ToString());
                 polipoConfig = polipoConfig.Replace("__POLIPO_BIND_IP__", configuration.shareOverLan ? "0.0.0.0" : "127.0.0.1");
-                FileManager.ByteArrayToFile(temppath + "/polipo.conf", System.Text.Encoding.UTF8.GetBytes(polipoConfig));
+                FileManager.ByteArrayToFile(runningPath + "/privoxy.conf", System.Text.Encoding.UTF8.GetBytes(polipoConfig));
 
                 _process = new Process();
                 // Configure the process using the StartInfo properties.
-                _process.StartInfo.FileName = temppath + "/ss_polipo.exe";
-                _process.StartInfo.Arguments = "-c \"" + temppath + "/polipo.conf\"";
+                _process.StartInfo.FileName = runningPath + "/ss_privoxy.exe";
+                _process.StartInfo.Arguments = " \"" + runningPath + "/privoxy.conf\"";
                 _process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
                 _process.StartInfo.UseShellExecute = true;
                 _process.StartInfo.CreateNoWindow = true;
@@ -74,6 +93,7 @@ namespace Shadowsocks.Controller
                 //_process.StartInfo.RedirectStandardError = true;
                 _process.Start();
             }
+            RefreshTrayArea();
         }
 
         public void Stop()
@@ -91,6 +111,7 @@ namespace Shadowsocks.Controller
                 }
                 _process = null;
             }
+            RefreshTrayArea();
         }
 
         private int GetFreePort()
@@ -121,6 +142,51 @@ namespace Shadowsocks.Controller
                 return defaultPort;
             }
             throw new Exception("No free port found.");
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct RECT
+        {
+            public int left;
+            public int top;
+            public int right;
+            public int bottom;
+        }
+        [DllImport("user32.dll")]
+        public static extern IntPtr FindWindow(string lpClassName, string lpWindowName);
+        [DllImport("user32.dll")]
+        public static extern IntPtr FindWindowEx(IntPtr hwndParent, IntPtr hwndChildAfter, string lpszClass, string lpszWindow);
+        [DllImport("user32.dll")]
+        public static extern bool GetClientRect(IntPtr hWnd, out RECT lpRect);
+        [DllImport("user32.dll")]
+        public static extern IntPtr SendMessage(IntPtr hWnd, uint msg, int wParam, int lParam);
+
+
+        public void RefreshTrayArea()
+        {
+            IntPtr systemTrayContainerHandle = FindWindow("Shell_TrayWnd", null);
+            IntPtr systemTrayHandle = FindWindowEx(systemTrayContainerHandle, IntPtr.Zero, "TrayNotifyWnd", null);
+            IntPtr sysPagerHandle = FindWindowEx(systemTrayHandle, IntPtr.Zero, "SysPager", null);
+            IntPtr notificationAreaHandle = FindWindowEx(sysPagerHandle, IntPtr.Zero, "ToolbarWindow32", "Notification Area");
+            if (notificationAreaHandle == IntPtr.Zero)
+            {
+                notificationAreaHandle = FindWindowEx(sysPagerHandle, IntPtr.Zero, "ToolbarWindow32", "User Promoted Notification Area");
+                IntPtr notifyIconOverflowWindowHandle = FindWindow("NotifyIconOverflowWindow", null);
+                IntPtr overflowNotificationAreaHandle = FindWindowEx(notifyIconOverflowWindowHandle, IntPtr.Zero, "ToolbarWindow32", "Overflow Notification Area");
+                RefreshTrayArea(overflowNotificationAreaHandle);
+            }
+            RefreshTrayArea(notificationAreaHandle);
+        }
+
+
+        private static void RefreshTrayArea(IntPtr windowHandle)
+        {
+            const uint wmMousemove = 0x0200;
+            RECT rect;
+            GetClientRect(windowHandle, out rect);
+            for (var x = 0; x < rect.right; x += 5)
+                for (var y = 0; y < rect.bottom; y += 5)
+                    SendMessage(windowHandle, wmMousemove, 0, (y << 16) + x);
         }
     }
 }
